neurips experiments complete

+++ data/yahooreviewcuisine/prepare.py
@@ -10,6 +10,7 @@ from enum import Enum
 from openai import OpenAI
 from pprint import pprint
 import time
+import hashlib
 
 
 class stella_400m:
@@ -289,16 +290,174 @@ class YelpEasyCFGTokenizer:
             raise ValueError(f"Unknown token: {last_token}")
         return mask
 
-def prepare_data(outputs_only: bool = False):
-    # 1. Load reviews.json
+
+def compare_with_llama_and_generations_2():
+    with open(os.path.join(output_dir, "llama_3_3_generations.json"), "r") as f:
+        data_llama = json.load(f)
+
     with open(os.path.join(output_dir, "with_generations_2.json"), "r") as f:
+        data_gpt = json.load(f)
+
+    print(f"Items of first item of data_gpt: {data_gpt[0].items()}")
+    for i in range(len(data_gpt)):
+        data_gpt[i]["generation"] = data_gpt[i]["generated"]
+        del data_gpt[i]["generated"]
+
+    # okay we need to re-order the things that   
+    # hash the inputs of both 
+    # sort the data_llama by the hash
+
+    # note that there are some entries in data_llama that are not populated, so we only look at those 
+    # that are populated with 'generation'
+
+    for i in range(len(data_llama)):
+        data_llama[i]["hash"] = hashlib.sha256(data_llama[i]["input"].encode()).hexdigest()
+    for i in range(len(data_gpt)):
+        data_gpt[i]["hash"] = hashlib.sha256(data_gpt[i]["input"].encode()).hexdigest()
+
+    data_gpt_map = {item["hash"]: item for item in data_gpt}
+
+    count_vs_gpt = 0
+    for i in range(len(data_llama)):
+        if "generation" not in data_llama[i]:
+            continue
+        if data_llama[i]["hash"] not in data_gpt_map:
+            print(f"Hash not in data_gpt: {data_llama[i]['hash']}")
+            continue
+        if data_llama[i]["generation"] == data_gpt_map[data_llama[i]["hash"]]["generation"]:
+            count_vs_gpt += 1
+    
+    print(f"Accuracy of founding: {count_vs_gpt / len(data_llama)}")
+
+    # now re-map
+    for i in range(len(data_llama)):
+        data_gpt_map[data_llama[i]["hash"]]["generation_llama"] = data_llama[i]["predicted"]
+        assert data_gpt_map[data_llama[i]["hash"]]["generation"] == data_llama[i]["generation"]
+
+    # convert to list, but make sure the ones with generation_llama are at the start
+    data_gpt_list = []
+    for i in range(len(data_gpt)):
+        if "generation_llama" in data_gpt[i]:
+            data_gpt_list.append(data_gpt[i])
+        
+    for i in range(len(data_gpt)):
+        if "generation_llama" not in data_gpt[i]:
+            data_gpt_list.append(data_gpt[i])
+
+    # let's make sure this is correct
+    last_index_with_generation_llama = None
+    for i in range(len(data_gpt_list)):
+        if "generation_llama" not in data_gpt_list[i]:
+            last_index_with_generation_llama = i
+            break
+    print(f"Last index with generation llama: {last_index_with_generation_llama}")
+            
+    
+    # now let's check that the accuracy for generations and generations_llama are about the same
+    count_correct_generations = 0
+    count_correct_generations_llama = 0
+    count_correct_generations_vs_llama = 0
+    last_index_with_generation_llama = None
+    for i in range(len(data_gpt_list)):
+        if "generation_llama" not in data_gpt_list[i]:
+            last_index_with_generation_llama = i
+            break
+        if data_gpt_list[i]["expected"] == data_gpt_list[i]["generation"]:
+            count_correct_generations += 1
+        if data_gpt_list[i]["expected"] == data_gpt_list[i]["generation_llama"]:
+            count_correct_generations_llama += 1
+        if data_gpt_list[i]["generation"] == data_gpt_list[i]["generation_llama"]:
+            count_correct_generations_vs_llama += 1
+    
+    print(f"Accuracy of generations: {count_correct_generations / (last_index_with_generation_llama + 1)}")
+    print(f"Accuracy of generations_llama: {count_correct_generations_llama / (last_index_with_generation_llama + 1)}")
+    print(f"Accuracy of generations vs generations_llama: {count_correct_generations_vs_llama / (last_index_with_generation_llama + 1)}")
+
+    # count the generation vs ground truth for the whole set
+    count_correct_generations_vs_ground_truth = 0
+    for i in range(len(data_gpt_list)):
+        if data_gpt_list[i]["generation"] == data_gpt_list[i]["expected"]:
+            count_correct_generations_vs_ground_truth += 1
+    print(f"Accuracy of generations vs ground truth: {count_correct_generations_vs_ground_truth / len(data_gpt_list)}")
+
+    with open(os.path.join(output_dir, "with_generations_with_llama.json"), "w") as f:
+        json.dump(data_gpt_list, f)
+
+
+def get_cost_per_example():
+    # get cost per example of gpt 4o mini
+    import tiktoken
+    file_name = "data/yahooreviewcuisine/with_generations_2.json"
+
+    with open(file_name, "r") as f:
         data = json.load(f)
 
-    inputs = [item["input"] for item in data]
-    outputs = [item["expected"] for item in data]
-    generation_outputs = [item["generated"] for item in data]
+    # get the cost per example
+    input_cost_per_token = 0.15 / 1000000
+    output_cost_per_token = 0.6 / 1000000
+    total_cost = 0
+    for item in data:
+        input_tokens = len(tiktoken.encoding_for_model("gpt-4o-mini").encode(item["input"]))
+        output_tokens = len(tiktoken.encoding_for_model("gpt-4o-mini").encode(item["generated"]))
+        total_cost += input_cost_per_token * input_tokens + output_cost_per_token * output_tokens
+    print(f"Total cost: {total_cost}")
+    print(f"Cost per example: {1000 * total_cost / len(data)}")
 
-    if not outputs_only:
+
+def prepare_data(keep_embedding_unchanged
+                 : bool = False):
+    
+    # 1. Load reviews.json
+    with open(os.path.join(output_dir, "with_generations_with_llama.json"), "r") as f:
+        data = json.load(f)
+
+    # let's first do checks
+    num_llama_generations = 0
+    num_gpt_generations = len(data)
+    count_correct_generations_vs_llama = 0
+    count_correct_generations_vs_llama_vs_ground_truth = 0
+    count_correct_generations_vs_ground_truth = 0
+    count_correct_generations = 0
+    for i in range(len(data)):
+        if "generation_llama" in data[i]:
+            num_llama_generations += 1
+            if data[i]["generation"] == data[i]["generation_llama"]:
+                count_correct_generations_vs_llama += 1
+            if data[i]["generation_llama"] == data[i]["expected"]:
+                count_correct_generations_vs_llama_vs_ground_truth += 1
+        if data[i]["generation"] == data[i]["expected"]:
+            count_correct_generations_vs_ground_truth += 1
+        if data[i]["generation"] == data[i]["generation"]:
+            count_correct_generations += 1
+    
+    print(f"Accuracy of generations vs llama: {count_correct_generations_vs_llama / num_llama_generations}")
+    print(f"Accuracy of generations vs llama vs ground truth: {count_correct_generations_vs_llama_vs_ground_truth / num_llama_generations}")
+    print(f"Accuracy of generations vs ground truth: {count_correct_generations_vs_ground_truth / num_gpt_generations}")
+    print(f"Accuracy of generations: {count_correct_generations / num_gpt_generations}")
+    print(f"Number of llama generations: {num_llama_generations}")
+    print(f"Number of gpt generations: {num_gpt_generations}")
+
+
+    print(f"Length of data: {len(data)}")
+
+    
+    for i in range(len(data)):
+        if "generation_llama" not in data[i]:
+            cutoff_with_last_generated = i
+            break
+    # print(f"Cutoff with last generated: {cutoff_with_last_generated}")
+    data_new_generated_only = data[:cutoff_with_last_generated]
+    generation_outputs = [item["generation"] for item in data_new_generated_only]
+
+    count_correct = 0
+    for i in range(len(generation_outputs)):
+        if generation_outputs[i] == data[i]["expected"]:
+            count_correct += 1
+    print(f"Accuracy: {count_correct / len(generation_outputs)}")
+
+    outputs = [item["expected"] for item in data]
+    if not keep_embedding_unchanged:
+        inputs = [item["input"] for item in data]
         # 2. Embed inputs
         time_start = time.time()
         stella = stella_400m()  # Replace with your actual Stella model
@@ -317,7 +476,9 @@ def prepare_data(outputs_only: bool = False):
             tokens += [tokenizer.pad_token_id] * (block_size - len(tokens))
         elif len(tokens) > block_size:
             raise ValueError(f"Output token length is greater than block size: {len(tokens)} > {block_size}")
-        
+        output_tokens.append(tokens)
+    
+    for i in range(len(generation_outputs)):
         # get the generated tokens
         generated_tokens = tokenizer.tokenize(generation_outputs[i])
         # Pad or truncate to block_size
@@ -325,28 +486,35 @@ def prepare_data(outputs_only: bool = False):
             generated_tokens += [tokenizer.pad_token_id] * (block_size - len(generated_tokens))
         elif len(generated_tokens) > block_size:
             raise ValueError(f"Generated token length is greater than block size: {len(generated_tokens)} > {block_size}")
-
-        output_tokens.append(tokens)
         generated_tokens_list.append(generated_tokens)
+
+    
     output_tokens = np.array(output_tokens, dtype=np.int16)  # shape: (num_samples, block_size)
     generated_tokens_list = np.array(generated_tokens_list, dtype=np.int16)  # shape: (num_samples, block_size)
 
     # check that all the lists are the same length
-    assert len(input_embeddings) == len(output_tokens) == len(generated_tokens_list)
-    # 4. Save to files
-    if not outputs_only:
-        np.save(os.path.join(output_dir, "input_embeddings_2.npy"), input_embeddings)
-    np.save(os.path.join(output_dir, "output_tokens_2.npy"), output_tokens)
-    np.save(os.path.join(output_dir, "generated_tokens_list_2.npy"), generated_tokens_list)
-    # Save tokenizer metadata for reference
-    with open(os.path.join(output_dir, "tokenizer_meta.pkl"), "wb") as f:
-        pickle.dump({
-            "vocab_size": tokenizer.get_vocab_size(),
-            "pad_token_id": tokenizer.pad_token_id,
-            "eos_token_id": tokenizer.eos_token_id,
-        }, f)
+    if not keep_embedding_unchanged:
+        assert len(input_embeddings) == len(output_tokens)
+        print(f"Input embeddings shape: {input_embeddings.shape}")
+    print(f"Output tokens shape: {output_tokens.shape}")
+    print(f"Generated tokens list shape: {generated_tokens_list.shape}")
+    
 
-    print(f"Saved {len(inputs)} samples to {output_dir}")
+    # 4. Save to files
+    if not keep_embedding_unchanged:
+        np.save(os.path.join(output_dir, "input_embeddings_3.npy"), input_embeddings)
+    np.save(os.path.join(output_dir, "output_tokens_3.npy"), output_tokens)
+    np.save(os.path.join(output_dir, "generated_tokens_list_3.npy"), generated_tokens_list)
+    # Save tokenizer metadata for reference
+    if not keep_embedding_unchanged:
+        with open(os.path.join(output_dir, "tokenizer_meta.pkl"), "wb") as f:
+            pickle.dump({
+                "vocab_size": tokenizer.get_vocab_size(),
+                "pad_token_id": tokenizer.pad_token_id,
+                "eos_token_id": tokenizer.eos_token_id,
+            }, f)
+
+    print(f"Saved {generated_tokens_list.shape[0]} samples to {output_dir}")
 
 in_memory_input_embeddings = None
 in_memory_ground_truth_tokens = None
@@ -364,16 +532,18 @@ def get_batch_yahoo(batch_size: int,  train_samples_limit, split: str = "train",
     global in_memory_input_embeddings, in_memory_ground_truth_tokens, in_memory_generated_tokens_list
 
     if in_memory_input_embeddings is None:
-        in_memory_input_embeddings = np.load(os.path.join(output_dir, "input_embeddings_2.npy"))
+        in_memory_input_embeddings = np.load(os.path.join(output_dir, "input_embeddings_3.npy"))
         print(f"Loaded input embeddings for {in_memory_input_embeddings.shape[0]} samples")
     if in_memory_ground_truth_tokens is None:
-        in_memory_ground_truth_tokens = np.load(os.path.join(output_dir, "output_tokens_2.npy")) 
+        in_memory_ground_truth_tokens = np.load(os.path.join(output_dir, "output_tokens_3.npy")) 
         print(f"Loaded output tokens for {in_memory_ground_truth_tokens.shape[0]} samples")
     if in_memory_generated_tokens_list is None:
-        in_memory_generated_tokens_list = np.load(os.path.join(output_dir, "generated_tokens_list_2.npy"))
+        in_memory_generated_tokens_list = np.load(os.path.join(output_dir, "generated_tokens_list_3.npy"))
         print(f"Loaded generated tokens list for {in_memory_generated_tokens_list.shape[0]} samples")
     
     num_samples = in_memory_input_embeddings.shape[0]
+    if split == "val":
+        num_samples = in_memory_generated_tokens_list.shape[0]
 
     assert not (train_samples_limit is None and split == "train")
     # batch size is only valid for training split (batches are done outside of this for val)
@@ -390,7 +560,10 @@ def get_batch_yahoo(batch_size: int,  train_samples_limit, split: str = "train",
     
     possible_choice_num = high - low
     if batch_size is not None and batch_size > possible_choice_num:
-        raise ValueError(f"Possible choice number is smaller than batch size: {possible_choice_num} < {batch_size}")
+        print(f"Possible choice number is smaller than batch size: {possible_choice_num} < {batch_size}")
+        batch_size = possible_choice_num
+    
+
     
     if split == "train":
         idx = rng.choice(possible_choice_num, size=batch_size, replace=False)
@@ -446,8 +619,9 @@ if __name__ == "__main__":
     # print("y_tokens[1]:", tokenizer.detokenize(y_tokens[1].tolist()))
 
     # print(get_chatgpt_completions())
-    prepare_data()
-
+    # prepare_data(keep_embedding_unchanged=False)
+    get_cost_per_example()
+    # compare_with_llama_and_generations_2()
     # data = json.load(open(os.path.join(output_dir, "with_generations_2.json"), "r"))
 
     # # count which have {restaurant(None)
